+++ tw-tasks-core/src/main/java/com/transferwise/tasks/domain/BaseTask.java
@@ -12,7 +12,6 @@ public class BaseTask implements IBaseTask {
   private String type;
   private long version;
   private int priority;
-  private String partitionKey;
 
   public ITaskVersionId getVersionId() {
     return new TaskVersionId(id, version);