+++ tw-tasks-core/src/main/java/com/transferwise/tasks/helpers/kafka/partitionkey/IPartitionKeyStrategy.java
@@ -7,6 +7,6 @@ import com.transferwise.tasks.domain.BaseTask;
  */
 public interface IPartitionKeyStrategy {
 
-  String getPartitionKey(BaseTask task);
+  String createPartitionKey(BaseTask task);
 
 }